requirements.txt
python-binance
python-dotenv
pandas
matplotlib
ta-lib
